requirements.txt
backoff==2.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
Deprecated==1.2.13
googleapis-common-protos==1.58.0
grpcio==1.51.3
idna==3.4
opentelemetry-api==1.16.0
opentelemetry-distro==0.37b0
opentelemetry-exporter-otlp==1.16.0
opentelemetry-exporter-otlp-proto-grpc==1.16.0
opentelemetry-exporter-otlp-proto-http==1.16.0
opentelemetry-instrumentation==0.37b0
opentelemetry-proto==1.16.0
opentelemetry-sdk==1.16.0
opentelemetry-semantic-conventions==0.37b0
protobuf==4.22.1
requests==2.28.2
typing_extensions==4.5.0
urllib3==1.26.15
wrapt==1.15.0
flask==2.3.2